requirements.txt
ncnn>=1.0.0
opencv-python>=4.8.0
pillow>=9.0.0
numpy>=1.24.0
reportlab>=3.6.0
PyQt5>=5.15.0
pyserial